line kick motion planner used (breaks all)
