+++ rj_gameplay/rj_gameplay/skill/line_kick.py
@@ -9,8 +9,7 @@ import time
 import stp.skill as skill
 import stp.role as role
 import stp.action as action
-from rj_gameplay.action import move
-from rj_gameplay.action import kick 
+from rj_gameplay.action import move, line_kick
 from stp.skill.action_behavior import ActionBehavior
 import stp.rc as rc
 
@@ -19,7 +18,7 @@ class ILineKick(skill.ISkill, ABC):
 
 class LineKick(ILineKick):
     """
-    Move through the ball and shoot.
+    A skill version of line kick so that actions don't have to be called in tactics
     """
 
     # role-based implementation
@@ -27,23 +26,20 @@ class LineKick(ILineKick):
         # self.robot = role.robot
     # role-blind implementation
     def __init__(self, robot: rc.Robot, world_state: rc.WorldState) -> None:
+        self.__name__ = "line kick skill"
         self.robot = robot
 
-        r_id = None
-        if robot:
-            r_id = self.robot.id
-        # move to ball
-        self.move = move.Move(r_id, world_state.ball.pos)
-        self.move_behavior = ActionBehavior('Move', self.move)
-        # kick once there
-        self.kick = kick.Kick(r_id)
-        self.kick_behavior = ActionBehavior('Kick', self.kick)
+        # setup line kick action
+        self.target_point = world_state.ball.pos 
+        if self.robot is not None:
+            self.line_kick = line_kick.LineKick(self.robot.id, self.target_point)
+        else:
+            self.line_kick = line_kick.LineKick(None, self.target_point)
 
-        # setup sequence of move -> kick
-        self.root = py_trees.composites.Sequence("Sequence")
-        self.root.add_children([self.move_behavior, self.kick_behavior])
+        # put into a tree
+        self.line_kick_behavior = ActionBehavior('LineKick', self.line_kick)
+        self.root = self.line_kick_behavior
         self.root.setup_with_descendants()
-        self.__name__ = "line kick skill"
 
     def tick(self, robot: rc.Robot, world_state: rc.WorldState) -> None:
         self.robot = robot
@@ -53,4 +49,4 @@ class LineKick(ILineKick):
 
     def is_done(self, world_state: rc.WorldState):
         # skill is done after move + kick
-        return self.kick.is_done(world_state)
+        return self.line_kick.is_done(world_state)
